+++ EssentialFeed/EssentialFeedTests/Feed Cache/LoadFeedImageDataFromCacheUseCaseTests.swift
@@ -13,16 +13,6 @@ class LoadFeedImageDataFromCacheUseCaseTests: XCTestCase {
 		XCTAssertTrue(store.receivedMessages.isEmpty)
 	}
 	
-	func test_saveImageDataForURL_requestsImageDataInsertionForURL() {
-		let (sut, store) = makeSUT()
-		let url = anyURL()
-		let data = anyData()
-		
-		sut.save(data, for: url) { _ in }
-		
-		XCTAssertEqual(store.receivedMessages, [.insert(data: data, for: url)])
-	}
-	
 	func test_loadImageDataFromURL_requestsStoredDataForURL() {
 		let (sut, store) = makeSUT()
 		let url = anyURL()
